chore(cleanup): remove dead code and relocate test-only seams
Systematic dead-code sweep driven by golang.org/x/tools/cmd/deadcode
(binary and -test modes) and staticcheck U1000. Net -2,148 lines.

Removed:
- The unused responsecache stream-reconstruction subsystem
  (stream_cache_chat.go, stream_cache_responses.go, and the dead half of
  stream_cache.go), superseded by verbatim raw-SSE replay, plus the
  tests that only exercised it.
- Dead thin wrappers, migrating their test call sites to the live
  variants: workflows.NewCompiler, dashboard.New,
  gateway.ResolveRequestModel, server resolveRequestModel,
  DetermineBatchExecutionSelection(WithAuthorizer),
  ensureTranslatedRequestWorkflow, doOpenAICompatibleFileIDRequest, and
  the batchrewrite RecordPreparation/CleanupSuperseded* family (the
  guardrails harness now mirrors production via RecordResult/CleanupFile).
- Dead functions together with their dead tests:
  observability.GetMetrics/HealthCheck, core.HasNonTextContent,
  core.NormalizeModelSelector, the DecodedBatchItemRequest accessors,
  BatchItemModelSelector, auditlog.DefaultConfig,
  conversationstore.WithUnboundedRetention, responsecache guardrails-hash
  and cache-skip wrappers, and the unused MockEmbedder test type.

Relocated production-dead but test-used seams into test files:
server NewHandler/newHandler/AuthMiddleware/WorkflowResolution(
WithResolver)/GetProviderType/handleStreamingResponse (new
seams_test.go), vecstore_map.go, NewRedisModelCacheWithStore,
failover.NewResolver, admin.WithGuardrailsRegistry, derivedEnvNames.

Deliberately kept: the ext package public extension API, the provider
NewWithHTTPClient constructor family (contract tests and cross-provider
delegation), and cross-package test seams (cache.MapStore,
llmclient.DefaultConfig, ResetMetrics, NewResponseCacheMiddlewareWithStore,
memory-store retention options).

deadcode -test now reports only the intentional ext API; staticcheck
U1000 (including tests) is clean; all packages pass, and contract/
integration suites compile under their build tags.
