+++ internal/batchrewrite/helpers.go
@@ -17,22 +17,6 @@ type FileDeleter interface {
 	DeleteFile(ctx context.Context, providerType, id string) (*core.FileDeleteResponse, error)
 }
 
-// FileRouter resolves a routed native file provider lazily.
-type FileRouter func() (core.NativeFileRoutableProvider, error)
-
-// RecordPreparation stores request-scoped rewrite metadata for persistence and
-// later cleanup.
-func RecordPreparation(ctx context.Context, original, rewritten *core.BatchRequest) {
-	if ctx == nil || original == nil || rewritten == nil {
-		return
-	}
-	metadata := core.GetBatchPreparationMetadata(ctx)
-	if metadata == nil {
-		return
-	}
-	metadata.RecordInputFileRewrite(original.InputFileID, rewritten.InputFileID)
-}
-
 // RecordResult stores rewrite metadata produced by an explicit batch preparer.
 func RecordResult(ctx context.Context, result *core.BatchRewriteResult) {
 	if ctx == nil || result == nil {
@@ -63,52 +47,6 @@ func CleanupFile(ctx context.Context, files FileDeleter, providerType, fileID, l
 		return false
 	}
 	return true
-}
-
-// CleanupFileFromRouter resolves the native file API only when there is a file
-// id to delete.
-func CleanupFileFromRouter(ctx context.Context, router FileRouter, providerType, fileID, logMessage string, attrs ...any) bool {
-	fileID = strings.TrimSpace(fileID)
-	if router == nil || fileID == "" {
-		return false
-	}
-	files, err := router()
-	if err != nil {
-		return false
-	}
-	return CleanupFile(ctx, files, providerType, fileID, logMessage, attrs...)
-}
-
-// CleanupSupersededFileFromRouter deletes a local rewrite artifact only when a
-// later rewrite has replaced it in the request-scoped batch metadata.
-func CleanupSupersededFileFromRouter(ctx context.Context, router FileRouter, providerType, fileID, logMessage string, attrs ...any) bool {
-	if !ShouldCleanupSupersededFile(ctx, fileID) {
-		return false
-	}
-	return CleanupFileFromRouter(ctx, router, providerType, fileID, logMessage, attrs...)
-}
-
-// CleanupSupersededFile deletes a local rewrite artifact only when a later
-// rewrite has replaced it in the request-scoped batch metadata.
-func CleanupSupersededFile(ctx context.Context, files FileDeleter, providerType, fileID, logMessage string, attrs ...any) bool {
-	if !ShouldCleanupSupersededFile(ctx, fileID) {
-		return false
-	}
-	return CleanupFile(ctx, files, providerType, fileID, logMessage, attrs...)
-}
-
-// ShouldCleanupSupersededFile reports whether fileID is a temporary rewrite
-// artifact that has been superseded by a later rewrite stage.
-func ShouldCleanupSupersededFile(ctx context.Context, fileID string) bool {
-	fileID = strings.TrimSpace(fileID)
-	if fileID == "" {
-		return false
-	}
-	metadata := core.GetBatchPreparationMetadata(ctx)
-	if metadata == nil {
-		return false
-	}
-	return strings.TrimSpace(metadata.RewrittenInputFileID) != fileID
 }
 
 // MergeEndpointHints returns a fresh map containing left hints overwritten by
