+++ internal/responsecache/handle_request_test.go
@@ -3,7 +3,6 @@ package responsecache
 import (
 	"bytes"
 	"context"
-	"encoding/json"
 	"errors"
 	"net/http"
 	"net/http/httptest"
@@ -1081,481 +1080,5 @@ func TestHandleRequest_InvalidStreamingBodySkipsExactCacheWrite(t *testing.T) {
 	}
 	if handlerCalls != 2 {
 		t.Fatalf("expected invalid stream to bypass cache on follow-up, got %d calls", handlerCalls)
-	}
-}
-
-func TestReconstructStreamingResponse_PreservesChatReasoningContent(t *testing.T) {
-	raw := []byte(
-		"data: {\"id\":\"chatcmpl-reasoning\",\"object\":\"chat.completion.chunk\",\"created\":1234567890,\"model\":\"claude-sonnet\",\"provider\":\"anthropic\",\"choices\":[{\"index\":0,\"delta\":{\"role\":\"assistant\",\"reasoning_content\":\"think first\"},\"finish_reason\":null}]}\n\n" +
-			"data: {\"id\":\"chatcmpl-reasoning\",\"object\":\"chat.completion.chunk\",\"created\":1234567890,\"model\":\"claude-sonnet\",\"provider\":\"anthropic\",\"choices\":[{\"index\":0,\"delta\":{\"content\":\"final answer\"},\"finish_reason\":\"stop\"}],\"usage\":{\"prompt_tokens\":10,\"completion_tokens\":4,\"total_tokens\":14}}\n\n" +
-			"data: [DONE]\n\n",
-	)
-
-	cached, ok := reconstructStreamingResponse("/v1/chat/completions", raw, streamResponseDefaults{
-		Model:    "claude-sonnet",
-		Provider: "anthropic",
-	})
-	if !ok {
-		t.Fatal("expected streamed chat response to reconstruct successfully")
-	}
-	if !bytes.Contains(cached, []byte(`"reasoning_content":"think first"`)) {
-		t.Fatalf("reconstructed chat response = %q, want reasoning_content preserved", string(cached))
-	}
-
-	replay, err := renderCachedChatStream([]byte(`{"model":"claude-sonnet","stream":true}`), cached)
-	if err != nil {
-		t.Fatalf("renderCachedChatStream() error = %v", err)
-	}
-	if !bytes.Contains(replay, []byte(`"reasoning_content":"think first"`)) {
-		t.Fatalf("cached chat replay = %q, want reasoning_content delta", string(replay))
-	}
-	if bytes.Contains(replay, []byte(`"usage"`)) {
-		t.Fatalf("cached chat replay without include_usage = %q, did not expect usage chunk", string(replay))
-	}
-
-	replayWithUsage, err := renderCachedChatStream([]byte(`{"model":"claude-sonnet","stream":true,"stream_options":{"include_usage":true}}`), cached)
-	if err != nil {
-		t.Fatalf("renderCachedChatStream(include_usage) error = %v", err)
-	}
-	if !bytes.Contains(replayWithUsage, []byte(`"usage"`)) {
-		t.Fatalf("cached chat replay with include_usage = %q, want usage chunk", string(replayWithUsage))
-	}
-}
-
-func TestRenderCachedChatStream_EmitsStandaloneUsageChunk(t *testing.T) {
-	raw := []byte(
-		"data: {\"id\":\"chatcmpl-usage\",\"object\":\"chat.completion.chunk\",\"created\":1234567890,\"model\":\"gpt-4o-mini\",\"provider\":\"openai\",\"choices\":[{\"index\":0,\"delta\":{\"role\":\"assistant\",\"content\":\"Hello\"},\"finish_reason\":null}]}\n\n" +
-			"data: {\"id\":\"chatcmpl-usage\",\"object\":\"chat.completion.chunk\",\"created\":1234567890,\"model\":\"gpt-4o-mini\",\"provider\":\"openai\",\"choices\":[{\"index\":0,\"delta\":{\"content\":\" world\"},\"finish_reason\":\"stop\"}],\"usage\":{\"prompt_tokens\":11,\"completion_tokens\":2,\"total_tokens\":13}}\n\n" +
-			"data: [DONE]\n\n",
-	)
-
-	cached, ok := reconstructStreamingResponse("/v1/chat/completions", raw, streamResponseDefaults{
-		Model:    "gpt-4o-mini",
-		Provider: "openai",
-	})
-	if !ok {
-		t.Fatal("expected streamed chat response to reconstruct successfully")
-	}
-
-	replay, err := renderCachedChatStream([]byte(`{"model":"gpt-4o-mini","stream":true,"stream_options":{"include_usage":true}}`), cached)
-	if err != nil {
-		t.Fatalf("renderCachedChatStream() error = %v", err)
-	}
-
-	var events []map[string]any
-	parseSSEJSONEvents(replay, func(event map[string]any) {
-		events = append(events, event)
-	})
-	if len(events) != 2 {
-		t.Fatalf("len(events) = %d, want 2 chat events before [DONE]", len(events))
-	}
-
-	firstChoices, ok := events[0]["choices"].([]any)
-	if !ok || len(firstChoices) != 1 {
-		t.Fatalf("first event choices = %#v, want len=1", events[0]["choices"])
-	}
-	if _, ok := events[0]["usage"]; ok {
-		t.Fatalf("first event should not carry usage, got %#v", events[0]["usage"])
-	}
-
-	secondChoices, ok := events[1]["choices"].([]any)
-	if !ok || len(secondChoices) != 0 {
-		t.Fatalf("usage event choices = %#v, want empty slice", events[1]["choices"])
-	}
-	usage, ok := events[1]["usage"].(map[string]any)
-	if !ok {
-		t.Fatalf("usage event usage = %#v, want object", events[1]["usage"])
-	}
-	if got, ok := jsonNumberToInt(usage["total_tokens"]); !ok || got != 13 {
-		t.Fatalf("usage.total_tokens = %#v, want 13", usage["total_tokens"])
-	}
-}
-
-func TestReconstructStreamingResponse_PreservesChatLogprobs(t *testing.T) {
-	raw := []byte(
-		"data: {\"id\":\"chatcmpl-logprobs\",\"object\":\"chat.completion.chunk\",\"created\":1234567890,\"model\":\"gpt-4o-mini\",\"provider\":\"openai\",\"choices\":[{\"index\":0,\"delta\":{\"role\":\"assistant\",\"content\":\"Hello\"},\"logprobs\":null,\"finish_reason\":null}]}\n\n" +
-			"data: {\"id\":\"chatcmpl-logprobs\",\"object\":\"chat.completion.chunk\",\"created\":1234567890,\"model\":\"gpt-4o-mini\",\"provider\":\"openai\",\"choices\":[{\"index\":0,\"delta\":{\"content\":\" world\"},\"logprobs\":null,\"finish_reason\":\"stop\"}]}\n\n" +
-			"data: [DONE]\n\n",
-	)
-
-	cached, ok := reconstructStreamingResponse("/v1/chat/completions", raw, streamResponseDefaults{
-		Model:    "gpt-4o-mini",
-		Provider: "openai",
-	})
-	if !ok {
-		t.Fatal("expected streamed chat response to reconstruct successfully")
-	}
-	if !bytes.Contains(cached, []byte(`"logprobs":null`)) {
-		t.Fatalf("reconstructed chat response = %q, want choice.logprobs preserved", string(cached))
-	}
-
-	replay, err := renderCachedChatStream([]byte(`{"model":"gpt-4o-mini","stream":true}`), cached)
-	if err != nil {
-		t.Fatalf("renderCachedChatStream() error = %v", err)
-	}
-	if !bytes.Contains(replay, []byte(`"logprobs":null`)) {
-		t.Fatalf("cached chat replay = %q, want choice.logprobs preserved", string(replay))
-	}
-}
-
-func TestReconstructStreamingResponse_PreservesResponsesReasoningText(t *testing.T) {
-	raw := []byte(
-		"event: response.created\n" +
-			"data: {\"type\":\"response.created\",\"response\":{\"id\":\"resp_reasoning_build\",\"object\":\"response\",\"created_at\":1234567890,\"model\":\"grok-4\",\"provider\":\"xai\",\"status\":\"in_progress\",\"output\":[]}}\n\n" +
-			"event: response.output_item.added\n" +
-			"data: {\"type\":\"response.output_item.added\",\"output_index\":0,\"item\":{\"id\":\"rs_1\",\"type\":\"reasoning\",\"status\":\"in_progress\",\"summary\":[]}}\n\n" +
-			"event: response.reasoning_text.delta\n" +
-			"data: {\"type\":\"response.reasoning_text.delta\",\"item_id\":\"rs_1\",\"output_index\":0,\"content_index\":0,\"delta\":\"step by\"}\n\n" +
-			"event: response.reasoning_text.delta\n" +
-			"data: {\"type\":\"response.reasoning_text.delta\",\"item_id\":\"rs_1\",\"output_index\":0,\"content_index\":1,\"delta\":\"step\"}\n\n" +
-			"event: response.output_item.done\n" +
-			"data: {\"type\":\"response.output_item.done\",\"output_index\":0,\"item\":{\"id\":\"rs_1\",\"type\":\"reasoning\",\"status\":\"completed\",\"summary\":[]}}\n\n" +
-			"event: response.completed\n" +
-			"data: {\"type\":\"response.completed\",\"response\":{\"id\":\"resp_reasoning_build\",\"object\":\"response\",\"created_at\":1234567890,\"model\":\"grok-4\",\"provider\":\"xai\",\"status\":\"completed\",\"output\":[{\"id\":\"rs_1\",\"type\":\"reasoning\",\"status\":\"completed\",\"summary\":[]}]}}\n\n" +
-			"data: [DONE]\n\n",
-	)
-
-	cached, ok := reconstructStreamingResponse("/v1/responses", raw, streamResponseDefaults{
-		Model:    "grok-4",
-		Provider: "xai",
-	})
-	if !ok {
-		t.Fatal("expected streamed responses payload to reconstruct successfully")
-	}
-
-	var response map[string]any
-	if err := json.Unmarshal(cached, &response); err != nil {
-		t.Fatalf("json.Unmarshal(cached) error = %v", err)
-	}
-	output, ok := response["output"].([]any)
-	if !ok || len(output) != 1 {
-		t.Fatalf("reconstructed output = %#v, want len=1", response["output"])
-	}
-	item, ok := output[0].(map[string]any)
-	if !ok {
-		t.Fatalf("reconstructed output[0] = %#v, want object", output[0])
-	}
-	if _, ok := item["content"]; ok {
-		t.Fatalf("reasoning item should not use content field, got %#v", item["content"])
-	}
-	summary, ok := item["summary"].([]any)
-	if !ok || len(summary) != 2 {
-		t.Fatalf("reconstructed reasoning summary = %#v, want len=2", item["summary"])
-	}
-	wantTexts := []string{"step by", "step"}
-	for i, wantText := range wantTexts {
-		part, ok := summary[i].(map[string]any)
-		if !ok {
-			t.Fatalf("reconstructed summary[%d] = %#v, want object", i, summary[i])
-		}
-		if got, _ := part["type"].(string); got != "reasoning_text" {
-			t.Fatalf("reconstructed summary part type = %q, want reasoning_text", got)
-		}
-		if got, _ := part["text"].(string); got != wantText {
-			t.Fatalf("reconstructed summary part text = %q, want %q", got, wantText)
-		}
-	}
-
-	replay, err := renderCachedResponsesStream([]byte(`{"model":"grok-4","stream":true}`), cached)
-	if err != nil {
-		t.Fatalf("renderCachedResponsesStream() error = %v", err)
-	}
-	var reasoningDeltas []map[string]any
-	parseSSEJSONEvents(replay, func(event map[string]any) {
-		if eventType, _ := event["type"].(string); eventType == "response.reasoning_text.delta" {
-			reasoningDeltas = append(reasoningDeltas, event)
-		}
-	})
-	if len(reasoningDeltas) != 2 {
-		t.Fatalf("cached responses replay = %q, want 2 reasoning_text delta events", string(replay))
-	}
-	for i, deltaEvent := range reasoningDeltas {
-		if got, _ := deltaEvent["delta"].(string); got != wantTexts[i] {
-			t.Fatalf("reasoning delta text = %q, want %q", got, wantTexts[i])
-		}
-		if got, _ := deltaEvent["item_id"].(string); got != "rs_1" {
-			t.Fatalf("reasoning delta item_id = %q, want rs_1", got)
-		}
-		if got, ok := jsonNumberToInt(deltaEvent["output_index"]); !ok || got != 0 {
-			t.Fatalf("reasoning delta output_index = %#v, want 0", deltaEvent["output_index"])
-		}
-		if got, ok := jsonNumberToInt(deltaEvent["content_index"]); !ok || got != i {
-			t.Fatalf("reasoning delta content_index = %#v, want %d", deltaEvent["content_index"], i)
-		}
-	}
-}
-
-func TestReconstructStreamingResponse_HonorsResponsesTextDeltaLocators(t *testing.T) {
-	raw := []byte(
-		"event: response.created\n" +
-			"data: {\"type\":\"response.created\",\"response\":{\"id\":\"resp_text_locator\",\"object\":\"response\",\"created_at\":1234567890,\"model\":\"gpt-4o-mini\",\"provider\":\"openai\",\"status\":\"in_progress\",\"output\":[{\"id\":\"rs_1\",\"type\":\"reasoning\",\"status\":\"in_progress\",\"summary\":[]}]}}\n\n" +
-			"event: response.output_text.delta\n" +
-			"data: {\"type\":\"response.output_text.delta\",\"item_id\":\"msg_1\",\"output_index\":1,\"content_index\":0,\"delta\":\"final\"}\n\n" +
-			"event: response.output_text.delta\n" +
-			"data: {\"type\":\"response.output_text.delta\",\"item_id\":\"msg_1\",\"output_index\":1,\"content_index\":1,\"delta\":\"answer\"}\n\n" +
-			"event: response.completed\n" +
-			"data: {\"type\":\"response.completed\",\"response\":{\"id\":\"resp_text_locator\",\"object\":\"response\",\"created_at\":1234567890,\"model\":\"gpt-4o-mini\",\"provider\":\"openai\",\"status\":\"completed\",\"output\":[{\"id\":\"rs_1\",\"type\":\"reasoning\",\"status\":\"completed\",\"summary\":[{\"type\":\"reasoning_text\",\"text\":\"step by step\"}]},{\"id\":\"msg_1\",\"type\":\"message\",\"role\":\"assistant\",\"status\":\"completed\",\"content\":[]}]}}\n\n" +
-			"data: [DONE]\n\n",
-	)
-
-	cached, ok := reconstructStreamingResponse("/v1/responses", raw, streamResponseDefaults{
-		Model:    "gpt-4o-mini",
-		Provider: "openai",
-	})
-	if !ok {
-		t.Fatal("expected streamed responses payload to reconstruct successfully")
-	}
-
-	var response map[string]any
-	if err := json.Unmarshal(cached, &response); err != nil {
-		t.Fatalf("json.Unmarshal(cached) error = %v", err)
-	}
-	output, ok := response["output"].([]any)
-	if !ok || len(output) != 2 {
-		t.Fatalf("reconstructed output = %#v, want len=2", response["output"])
-	}
-	reasoningItem, ok := output[0].(map[string]any)
-	if !ok {
-		t.Fatalf("reconstructed output[0] = %#v, want object", output[0])
-	}
-	if _, ok := reasoningItem["content"]; ok {
-		t.Fatalf("reasoning item should not use content field, got %#v", reasoningItem["content"])
-	}
-	messageItem, ok := output[1].(map[string]any)
-	if !ok {
-		t.Fatalf("reconstructed output[1] = %#v, want object", output[1])
-	}
-	content, ok := messageItem["content"].([]any)
-	if !ok || len(content) != 2 {
-		t.Fatalf("reconstructed message content = %#v, want len=2", messageItem["content"])
-	}
-	wantTexts := []string{"final", "answer"}
-	for i, wantText := range wantTexts {
-		messagePart, ok := content[i].(map[string]any)
-		if !ok {
-			t.Fatalf("reconstructed message content[%d] = %#v, want object", i, content[i])
-		}
-		if got, _ := messagePart["text"].(string); got != wantText {
-			t.Fatalf("reconstructed message text = %q, want %q", got, wantText)
-		}
-	}
-
-	replay, err := renderCachedResponsesStream([]byte(`{"model":"gpt-4o-mini","stream":true}`), cached)
-	if err != nil {
-		t.Fatalf("renderCachedResponsesStream() error = %v", err)
-	}
-	var textDeltas []map[string]any
-	parseSSEJSONEvents(replay, func(event map[string]any) {
-		if eventType, _ := event["type"].(string); eventType == "response.output_text.delta" {
-			textDeltas = append(textDeltas, event)
-		}
-	})
-	if len(textDeltas) != 2 {
-		t.Fatalf("cached responses replay = %q, want 2 output_text delta events", string(replay))
-	}
-	for i, deltaEvent := range textDeltas {
-		if got, _ := deltaEvent["delta"].(string); got != wantTexts[i] {
-			t.Fatalf("text delta text = %q, want %q", got, wantTexts[i])
-		}
-		if got, _ := deltaEvent["item_id"].(string); got != "msg_1" {
-			t.Fatalf("text delta item_id = %q, want msg_1", got)
-		}
-		if got, ok := jsonNumberToInt(deltaEvent["output_index"]); !ok || got != 1 {
-			t.Fatalf("text delta output_index = %#v, want 1", deltaEvent["output_index"])
-		}
-		if got, ok := jsonNumberToInt(deltaEvent["content_index"]); !ok || got != i {
-			t.Fatalf("text delta content_index = %#v, want %d", deltaEvent["content_index"], i)
-		}
-	}
-}
-
-func TestRenderCachedResponsesStream_PreservesReasoningTextDeltas(t *testing.T) {
-	cached := []byte(`{
-		"id":"resp_reasoning",
-		"object":"response",
-		"created_at":1234567890,
-		"model":"grok-4",
-		"provider":"xai",
-		"status":"completed",
-		"output":[
-			{
-				"id":"rs_1",
-				"type":"reasoning",
-				"status":"completed",
-				"summary":[{"type":"reasoning_text","text":"step by step"}]
-			},
-			{
-				"id":"msg_1",
-				"type":"message",
-				"role":"assistant",
-				"status":"completed",
-				"content":[{"type":"output_text","text":"final answer"}]
-			}
-		]
-	}`)
-
-	replay, err := renderCachedResponsesStream([]byte(`{"model":"grok-4","stream":true}`), cached)
-	if err != nil {
-		t.Fatalf("renderCachedResponsesStream() error = %v", err)
-	}
-	if !bytes.Contains(replay, []byte("event: response.reasoning_text.delta")) {
-		t.Fatalf("cached responses replay = %q, want reasoning_text delta event", string(replay))
-	}
-	if !bytes.Contains(replay, []byte("step by step")) {
-		t.Fatalf("cached responses replay = %q, want reasoning delta text", string(replay))
-	}
-	if !bytes.Contains(replay, []byte("event: response.output_text.delta")) {
-		t.Fatalf("cached responses replay = %q, want output_text delta event", string(replay))
-	}
-}
-
-func TestRenderCachedResponsesStream_FunctionCallAddedItemOmitsArguments(t *testing.T) {
-	cached := []byte(`{
-		"id":"resp_function_call",
-		"object":"response",
-		"created_at":1234567890,
-		"model":"gpt-4o-mini",
-		"provider":"openai",
-		"status":"completed",
-		"output":[
-			{
-				"id":"fc_1",
-				"type":"function_call",
-				"status":"completed",
-				"call_id":"call_1",
-				"name":"lookup_weather",
-				"arguments":"{\"city\":\"Warsaw\"}"
-			}
-		]
-	}`)
-
-	replay, err := renderCachedResponsesStream([]byte(`{"model":"gpt-4o-mini","stream":true}`), cached)
-	if err != nil {
-		t.Fatalf("renderCachedResponsesStream() error = %v", err)
-	}
-
-	var addedItem map[string]any
-	var argDelta map[string]any
-	var argDone map[string]any
-	parseSSEJSONEvents(replay, func(event map[string]any) {
-		switch eventType, _ := event["type"].(string); eventType {
-		case "response.output_item.added":
-			addedItem, _ = event["item"].(map[string]any)
-		case "response.function_call_arguments.delta":
-			argDelta = event
-		case "response.function_call_arguments.done":
-			argDone = event
-		}
-	})
-
-	if addedItem == nil {
-		t.Fatalf("cached responses replay = %q, want output_item.added event", string(replay))
-	}
-	if _, ok := addedItem["arguments"]; ok {
-		t.Fatalf("added item arguments = %#v, want omitted", addedItem["arguments"])
-	}
-	if argDelta == nil || argDone == nil {
-		t.Fatalf("cached responses replay = %q, want function_call_arguments delta and done events", string(replay))
-	}
-	if got, _ := argDelta["delta"].(string); got != `{"city":"Warsaw"}` {
-		t.Fatalf("arguments delta = %q, want full arguments", got)
-	}
-	if got, _ := argDone["arguments"].(string); got != `{"city":"Warsaw"}` {
-		t.Fatalf("arguments done = %q, want full arguments", got)
-	}
-}
-
-func TestReconstructStreamingResponse_PreservesResponsesTerminalEvents(t *testing.T) {
-	tests := []struct {
-		name             string
-		eventName        string
-		status           string
-		terminalResponse string
-		assertTerminal   func(*testing.T, map[string]any)
-	}{
-		{
-			name:             "failed",
-			eventName:        "response.failed",
-			status:           "failed",
-			terminalResponse: `{"id":"resp_failed","object":"response","created_at":1234567890,"model":"gpt-4o-mini","provider":"openai","status":"failed","error":{"code":"boom","message":"upstream failed"},"metadata":{"trace":"abc"},"output":[]}`,
-			assertTerminal: func(t *testing.T, response map[string]any) {
-				t.Helper()
-				errMap, ok := response["error"].(map[string]any)
-				if !ok {
-					t.Fatalf("terminal response error = %#v, want object", response["error"])
-				}
-				if got, _ := errMap["code"].(string); got != "boom" {
-					t.Fatalf("terminal response error.code = %q, want boom", got)
-				}
-			},
-		},
-		{
-			name:             "incomplete",
-			eventName:        "response.incomplete",
-			status:           "incomplete",
-			terminalResponse: `{"id":"resp_incomplete","object":"response","created_at":1234567890,"model":"gpt-4o-mini","provider":"openai","status":"incomplete","incomplete_details":{"reason":"max_output_tokens"},"metadata":{"trace":"def"},"output":[]}`,
-			assertTerminal: func(t *testing.T, response map[string]any) {
-				t.Helper()
-				details, ok := response["incomplete_details"].(map[string]any)
-				if !ok {
-					t.Fatalf("terminal response incomplete_details = %#v, want object", response["incomplete_details"])
-				}
-				if got, _ := details["reason"].(string); got != "max_output_tokens" {
-					t.Fatalf("terminal response incomplete_details.reason = %q, want max_output_tokens", got)
-				}
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			raw := []byte(
-				"event: response.created\n" +
-					"data: {\"type\":\"response.created\",\"response\":{\"id\":\"resp_terminal\",\"object\":\"response\",\"created_at\":1234567890,\"model\":\"gpt-4o-mini\",\"provider\":\"openai\",\"status\":\"in_progress\",\"output\":[]}}\n\n" +
-					"event: " + tt.eventName + "\n" +
-					"data: {\"type\":\"" + tt.eventName + "\",\"response\":" + tt.terminalResponse + "}\n\n" +
-					"data: [DONE]\n\n",
-			)
-
-			cached, ok := reconstructStreamingResponse("/v1/responses", raw, streamResponseDefaults{
-				Model:    "gpt-4o-mini",
-				Provider: "openai",
-			})
-			if !ok {
-				t.Fatal("expected streamed responses payload to reconstruct successfully")
-			}
-
-			var cachedResponse map[string]any
-			if err := json.Unmarshal(cached, &cachedResponse); err != nil {
-				t.Fatalf("json.Unmarshal(cached) error = %v", err)
-			}
-			if got, _ := cachedResponse["status"].(string); got != tt.status {
-				t.Fatalf("cached response status = %q, want %q", got, tt.status)
-			}
-			tt.assertTerminal(t, cachedResponse)
-
-			replay, err := renderCachedResponsesStream([]byte(`{"model":"gpt-4o-mini","stream":true}`), cached)
-			if err != nil {
-				t.Fatalf("renderCachedResponsesStream() error = %v", err)
-			}
-
-			var terminalEvent map[string]any
-			parseSSEJSONEvents(replay, func(event map[string]any) {
-				if eventType, _ := event["type"].(string); eventType == tt.eventName {
-					terminalEvent = event
-				}
-			})
-			if terminalEvent == nil {
-				t.Fatalf("cached responses replay = %q, want terminal event %s", string(replay), tt.eventName)
-			}
-			response, ok := terminalEvent["response"].(map[string]any)
-			if !ok {
-				t.Fatalf("terminal event response = %#v, want object", terminalEvent["response"])
-			}
-			if got, _ := response["status"].(string); got != tt.status {
-				t.Fatalf("terminal event status = %q, want %q", got, tt.status)
-			}
-			tt.assertTerminal(t, response)
-		})
 	}
 }
